chore: Add missing workflow permissions (#37)
Add workflow permissions. See this example for more on this:
https://github.com/coopnorge/helloworld/security/code-scanning/19

If you check the security tab of the repo, you may see security issues
titled "Workflow does not contain permissions".
These issues is what this PR attempts to solve.

This PR was automatically generated by a codemod implemented in the
following PR:
- https://github.com/coopnorge/codemod/pull/136