refactor: Return path and cleanup in WriteTempFile (#58)
This refactor reduces the scope of the function. Instead of returning an
`*os.File` and allow the user to do what they want the function now just
returns the path to the file as the intent is to write a file and pass
the path
to a CLI tool.

To help the user cleanup after the file is no longer needed a cleanup
function
is returned.