build: Revert to Go 1.23
Requiring Go 1.24 will cause every app/library that wants to use this library to also require 1.24. This is fine for go-apps, but libraries still want to stay on 1.23, until 1.25 is released.

The only real requirement we have on 1.24 in this repo, is the `t.Chdir` function, which we can copy the behviour of, until Go 1.25 is released.