chore: Use go version from go.mod
This change removes the workaround of setting the Go version in the workflow.
The original and intended design was to read to Go version from the root
`go.mod` file. With the Engineering guidelines now requiring repos to be at Go
1.25 (with some exceptions allowing Go 1.24) all repositories after updating
their Go version now have the required tool chain to use the `tool` directive.