feat: implement a detection of changes for a specific language
This will detect changes compared to the main branch for a specific
language. This can be used in CI to detect if CI is neccesary for a
language.
This feature could be futher extended to have local ci that only runs
for certain changes but that is beyond scope for now.

Signed-off-by: Atze de Vries <atze.wiebe.de.vries@coop.no>