feat: implement a detection of changes for a specific language (#270)
This will detect changes compared to the main branch for a specific
language. This can be used in CI to detect if CI is necesary for a
language. This feature could be further extended to have local CC that
only runs
for certain changes but that is beyond scope for now.

It will also introduce workflows related to the public target. So if you
use `goapp`
you use a `goapp.yaml` and if you use `golib` you get `golib.yaml`. Only
`goapp.yaml` is implemented now. 

The `goapp` workflow will now also do the full Terraform CI including
spacelift checks.

The documentation is also updated with now to call the workflow for
goapp.

---------

Signed-off-by: Atze de Vries <atze.wiebe.de.vries@coop.no>
Co-authored-by: Niels Henrik Hagen <nhhagen@gmail.com>